Updated pointer size for ptr
it was doing some funky memory stuff but casting to an int first, it should be casting to a intptr_t

+++ linuxdoom-1.10/r_draw.c
@@ -41,6 +41,8 @@
 // State.
 #include "doomstat.h"
 
+#include "stdint.h"
+
 
 // ?
 #define MAXWIDTH			1120
@@ -460,7 +462,7 @@ void R_InitTranslationTables (void)
     int		i;
 	
     translationtables = Z_Malloc (256*3+255, PU_STATIC, 0);
-    translationtables = (byte *)(( (int)translationtables + 255 )& ~255);
+    translationtables = (byte *)(( (intptr_t)translationtables + 255 )& ~255);
     
     // translate just the 16 green colors
     for (i=0 ; i<256 ; i++)
